requirements.txt
Django==1.8
django-ipware=1.0.0
psycopg2==2.6
